correcion de bugs encontrados en produccion

+++ backend/src/main/java/com/vflores/pos/routesales/api/dto/RouteSaleDetailResponse.java
@@ -6,7 +6,7 @@ import java.util.UUID;
 public record RouteSaleDetailResponse(
         UUID productId,
         String productName,
-        Integer quantity,
+        BigDecimal quantity,
         BigDecimal price,
         BigDecimal subtotal
 ) {
